pyproject.toml
[project]
name = "productizing-deepseek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "huggingface-hub>=0.28.1",
    "ipython>=8.32.0",
    "modal>=0.73.51",
    "openai>=1.63.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.6",
    "together>=1.4.1",
]
